Move CLI smoke-test helpers to scripts/test-cli/ subfolder
Consolidates the duplicated bits across all five scripts/test-cli*.ts
harnesses into a small role-split module:

  scripts/test-cli/
    binaries.ts    — LOADER, BARE, REPL, TESTRUNNER, BUNDLER, BENCHRUNNER
    assertions.ts  — containsLine, normalizeLineEndings
    tmpdir.ts      — mkdtemp, makeTmpFactory, clean

Each consumer now imports only what it needs. The test-cli-helpers.ts
single-file shim added in the previous commit is removed; tmpdir-prefix
binding stays close to each harness via either makeTmpFactory("...")
(when one prefix per file) or mkdtemp("...") inline (when prefixes
differ per block, like test-cli.ts).
