+++ scripts/test-cli-config.ts
@@ -9,26 +9,23 @@
 
 import { $ } from "bun";
 import {
-  mkdtempSync,
   writeFileSync,
   readFileSync,
   existsSync,
-  rmSync,
   mkdirSync,
 } from "fs";
 import { join, resolve } from "path";
-import { tmpdir } from "os";
-import { containsLine } from "./test-cli-helpers";
+import {
+  LOADER,
+  REPL,
+  TESTRUNNER,
+  BUNDLER,
+  BENCHRUNNER,
+} from "./test-cli/binaries";
+import { containsLine } from "./test-cli/assertions";
+import { makeTmpFactory, clean } from "./test-cli/tmpdir";
 
-const ext = process.platform === "win32" ? ".exe" : "";
-const LOADER = `./build/GocciaScriptLoader${ext}`;
-const REPL = `./build/GocciaREPL${ext}`;
-const TESTRUNNER = `./build/GocciaTestRunner${ext}`;
-const BUNDLER = `./build/GocciaBundler${ext}`;
-const BENCHRUNNER = `./build/GocciaBenchmarkRunner${ext}`;
-
-const makeTmp = () => mkdtempSync(join(tmpdir(), "goccia-cfg-"));
-const clean = (d: string) => rmSync(d, { recursive: true, force: true });
+const makeTmp = makeTmpFactory("goccia-cfg-");
 
 /** Runs a binary with cwd set to a specific directory. */
 function runCwd(
